+++ platformio/__init__.py
@@ -46,7 +46,7 @@ __core_packages__ = {
     "contrib-piohome": "~3.4.2",
     "contrib-pioremote": "~1.0.0",
     "tool-scons": "~4.40502.0",
-    "tool-cppcheck": "~1.270.0",
+    "tool-cppcheck": "~1.21100.0",
     "tool-clangtidy": "~1.150005.0",
     "tool-pvs-studio": "~7.18.0",
 }